+++ app/db/base.py
@@ -1,0 +1,6 @@
+"""Base class for all database classes
+"""
+
+from sqlalchemy.ext.declarative import declarative_base
+
+BASE = declarative_base()
